backend/requirements.txt
# Core dependencies
fastapi[standard]>=0.115.6
uvicorn>=0.34.0
pydantic>=2.10.3
pydantic-settings>=2.7.0

# OpenAI
openai>=1.58.1

# Async HTTP
httpx>=0.28.1

# Templating
Jinja2>=3.1.0

# Environment
python-dotenv>=1.0.1

# Rate limiting with Upstash Redis
upstash-redis>=1.0.0

aiofiles>=23.2.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
celery>=5.3.0
redis>=5.0.0
upstash-redis>=1.0.0
PyJWT>=2.8.0
python-jose>=3.3.0
websockets>=14.0.0
boto3>=1.34.0
supabase>=2.0.0
bleach>=6.1.0  # SECURITY: HTML sanitization for XSS prevention
python-magic>=0.4.27  # SECURITY: File content validation for MIME bypass prevention
